refactor[rust]: keep imports private in polars-lazy::logical_plan::optimize
