+++ polars/polars-lazy/src/physical_plan/executors/scan/csv.rs
@@ -1,5 +1,4 @@
 use super::*;
-use crate::prelude::file_caching::FileFingerPrint;
 
 pub struct CsvExec {
     pub path: PathBuf,
