+++ polars/polars-lazy/src/physical_plan/executors/scan/parquet.rs
@@ -1,5 +1,4 @@
 use super::*;
-use crate::prelude::file_caching::FileFingerPrint;
 
 pub struct ParquetExec {
     path: PathBuf,
