+++ polars/polars-lazy/src/physical_plan/file_cache.rs
@@ -2,7 +2,8 @@ use std::sync::Mutex;
 
 use polars_core::prelude::*;
 
-use crate::prelude::file_caching::FileFingerPrint;
+#[cfg(any(feature = "parquet", feature = "csv-file", feature = "ipc"))]
+use crate::logical_plan::optimizer::file_caching::FileFingerPrint;
 use crate::prelude::*;
 
 #[derive(Clone)]
