+++ polars/polars-lazy/src/utils.rs
@@ -371,61 +371,6 @@ where
     single_pred.expect("an empty iterator was passed")
 }
 
-#[cfg(test)]
-pub(crate) mod test {
-    use polars_core::prelude::*;
-
-    use crate::prelude::stack_opt::{OptimizationRule, StackOptimizer};
-    use crate::prelude::*;
-
-    pub fn optimize_lp(lp: LogicalPlan, rules: &mut [Box<dyn OptimizationRule>]) -> LogicalPlan {
-        // initialize arena's
-        let mut expr_arena = Arena::with_capacity(64);
-        let mut lp_arena = Arena::with_capacity(32);
-        let root = to_alp(lp, &mut expr_arena, &mut lp_arena).unwrap();
-
-        let opt = StackOptimizer {};
-        let lp_top = opt.optimize_loop(rules, &mut expr_arena, &mut lp_arena, root);
-        node_to_lp(lp_top, &mut expr_arena, &mut lp_arena)
-    }
-
-    pub fn optimize_expr(
-        expr: Expr,
-        schema: Schema,
-        rules: &mut [Box<dyn OptimizationRule>],
-    ) -> Expr {
-        // initialize arena's
-        let mut expr_arena = Arena::with_capacity(64);
-        let mut lp_arena = Arena::with_capacity(32);
-        let schema = Arc::new(schema);
-
-        // dummy input needed to put the schema
-        let input = Box::new(LogicalPlan::Projection {
-            expr: vec![],
-            input: Box::new(Default::default()),
-            schema: schema.clone(),
-        });
-
-        let lp = LogicalPlan::Projection {
-            expr: vec![expr],
-            input,
-            schema,
-        };
-
-        let root = to_alp(lp, &mut expr_arena, &mut lp_arena).unwrap();
-
-        let opt = StackOptimizer {};
-        let lp_top = opt.optimize_loop(rules, &mut expr_arena, &mut lp_arena, root);
-        if let LogicalPlan::Projection { mut expr, .. } =
-            node_to_lp(lp_top, &mut expr_arena, &mut lp_arena)
-        {
-            expr.pop().unwrap()
-        } else {
-            unreachable!()
-        }
-    }
-}
-
 /// Get a set of the data source paths in this LogicalPlan
 pub(crate) fn agg_source_paths(
     root_lp: Node,
